requirements.txt
fastapi
uvicorn
ultralytics
Pillow
torch
numpy
python-multipart
dill
starlette
itsdangerous